+++ lib/pmdtester/builders/pmd_report_builder.rb
@@ -24,46 +24,52 @@ module PmdTester
     def get_pmd_binary_file
       logger.info "#{@pmd_branch_name}: Start packaging PMD"
       Dir.chdir(@local_git_repo) do
-        current_head_sha = Cmd.execute('git rev-parse HEAD')
-        current_branch_sha = Cmd.execute("git rev-parse #{@pmd_branch_name}")
+        build_branch_sha = Cmd.execute("git rev-parse #{@pmd_branch_name}^{commit}")
 
-        @pmd_version = determine_pmd_version
+        checkout_build_branch # needs a clean working tree, otherwise fails
 
-        # in case we are already on the correct branch
-        # and a binary zip already exists...
-        if current_head_sha == current_branch_sha &&
-           File.exist?("pmd-dist/target/pmd-bin-#{@pmd_version}.zip")
-          logger.warn "#{@pmd_branch_name}: Skipping packaging - zip for " \
-                      "#{@pmd_version} already exists"
+        raise "Wrong branch #{get_last_commit_sha}" unless build_branch_sha == get_last_commit_sha
+
+        logger.debug "#{@pmd_branch_name}: PMD Version is #{@pmd_version} " \
+                     "(sha=#{build_branch_sha})"
+        distro_path = saved_distro_path(build_branch_sha)
+        if File.directory?(distro_path)
+          logger.info "#{@pmd_branch_name}: Skipping packaging - saved version exists " \
+                      " in #{distro_path}"
         else
-          build_pmd
+          build_pmd(into_dir: distro_path)
         end
 
-        @pmd_branch_details.branch_last_sha = get_last_commit_sha
+        # we're still on the build branch
+        @pmd_branch_details.branch_last_sha = build_branch_sha
         @pmd_branch_details.branch_last_message = get_last_commit_message
-
-        logger.info "#{@pmd_branch_name}: Extracting the zip"
-        unzip_cmd = "unzip -qo pmd-dist/target/pmd-bin-#{@pmd_version}.zip -d #{@pwd}/target"
-        Cmd.execute(unzip_cmd)
       end
       logger.info "#{@pmd_branch_name}: Packaging PMD completed"
     end
 
-    def build_pmd
-      logger.info "#{@pmd_branch_name}: Checking out the branch"
-      checkout_cmd = "git checkout #{@pmd_branch_name}"
-      Cmd.execute(checkout_cmd)
+    # builds pmd on currently checked out branch
+    def build_pmd(into_dir:)
+      # in CI there might have been a build performed already. In that case
+      # we reuse the build result, otherwise we build PMD freshly
+      pmd_dist_target = "pmd-dist/target/pmd-bin-#{@pmd_version}.zip"
+      if File.exist?("#{@local_git_repo}/#{pmd_dist_target}")
+        # that's a warning, because we don't know, whether this build really
+        # belongs to the current branch or whether it's from a previous branch.
+        # In CI, that's not a problem, because the workspace is always fresh.
+        logger.warn "#{@pmd_branch_name}: Reusing already existing #{pmd_dist_target}"
+      else
+        logger.info "#{@pmd_branch_name}: Building PMD #{@pmd_version}..."
+        package_cmd = './mvnw clean package' \
+                      ' -Dmaven.test.skip=true' \
+                      ' -Dmaven.javadoc.skip=true' \
+                      ' -Dmaven.source.skip=true' \
+                      ' -Dcheckstyle.skip=true'
+        Cmd.execute(package_cmd)
+      end
 
-      # determine the version again - it might be different in the other branch
-      @pmd_version = determine_pmd_version
-
-      logger.info "#{@pmd_branch_name}: Building PMD #{@pmd_version}..."
-      package_cmd = './mvnw clean package' \
-                    ' -Dmaven.test.skip=true' \
-                    ' -Dmaven.javadoc.skip=true' \
-                    ' -Dmaven.source.skip=true' \
-                    ' -Dcheckstyle.skip=true'
-      Cmd.execute(package_cmd)
+      logger.info "#{@pmd_branch_name}: Extracting the zip"
+      Cmd.execute("unzip -qo #{pmd_dist_target} -d pmd-dist/target/exploded")
+      Cmd.execute("mv pmd-dist/target/exploded/pmd-bin-#{@pmd_version} #{into_dir}")
     end
 
     def determine_pmd_version
@@ -73,7 +79,7 @@ module PmdTester
     end
 
     def get_last_commit_sha
-      get_last_commit_sha_cmd = 'git rev-parse HEAD'
+      get_last_commit_sha_cmd = 'git rev-parse HEAD^{commit}'
       Cmd.execute(get_last_commit_sha_cmd)
     end
 
@@ -84,7 +90,7 @@ module PmdTester
 
     def generate_pmd_report(project)
       error_recovery_options = @error_recovery ? 'PMD_JAVA_OPTS="-Dpmd.error_recovery -ea" ' : ''
-      run_path = "target/pmd-bin-#{@pmd_version}/bin/run.sh"
+      run_path = "#{saved_distro_path(@pmd_branch_details.branch_last_sha)}/bin/run.sh"
       pmd_cmd = "#{error_recovery_options}" \
                 "#{run_path} pmd -d #{project.local_source_path} -f xml " \
                 "-R #{project.get_config_path(@pmd_branch_name)} " \
@@ -148,6 +154,40 @@ module PmdTester
       @project_builder.build_projects
       get_pmd_binary_file
       generate_pmd_reports
+    end
+
+    private
+
+    def checkout_build_branch
+      logger.info "#{@pmd_branch_name}: Checking out the branch"
+      # note that this would fail if the tree is dirty
+      Cmd.execute("git checkout #{@pmd_branch_name}")
+
+      # determine the version
+      @pmd_version = determine_pmd_version
+
+      return unless wd_has_dirty_git_changes
+
+      # working dir is dirty....
+      # we don't allow this because we need the SHA to address the zip file
+      logger.error "#{@pmd_branch_name}: Won\'t build without a clean working tree, " \
+                   'commit your changes'
+    end
+
+    def work_dir
+      "#{@pwd}/target"
+    end
+
+    # path to the unzipped distribution
+    # e.g. <cwd>/pmd-bin-<version>-<branch>-<sha>
+    def saved_distro_path(build_sha)
+      "#{work_dir}/pmd-bin-#{@pmd_version}" \
+      "-#{PmdBranchDetail.branch_filename(@pmd_branch_name)}" \
+      "-#{build_sha}"
+    end
+
+    def wd_has_dirty_git_changes
+      !Cmd.execute('git status --porcelain').empty?
     end
   end
 end