+++ test/test_pmd_report_builder.rb
@@ -4,6 +4,24 @@ require 'test_helper'
 
 # Unit test class for PmdReportBuilder
 class TestPmdReportBuilder < Test::Unit::TestCase
+  def setup
+    # pmd version that is simulated in tests when pmd should be built
+    @pmd_version = '6.10.0-SNAPSHOT'
+  end
+
+  def teardown
+    pmd_repo = 'target/repositories/pmd'
+    # pre-built PMD binary
+    pmd_binary = "#{pmd_repo}/pmd-dist/target/pmd-bin-#{@pmd_version}.zip"
+    File.unlink pmd_binary if File.exist? pmd_binary
+
+    # only deleting empty directories in order to leave pmd_repo intact
+    # for local dev environment, where a local pmd clone might already exist
+    ["#{pmd_repo}/pmd-dist/target", "#{pmd_repo}/pmd-dist", pmd_repo].each do |d|
+      Dir.unlink(d) if Dir.exist?(d) && Dir.empty?(d)
+    end
+  end
+
   def test_build_skip
     projects = []
     argv = %w[-r target/repositories/pmd -b master -p pmd_releases/6.1.0
@@ -11,7 +29,32 @@ class TestPmdReportBuilder < Test::Unit::TestCase
     options = PmdTester::Options.new(argv)
 
     record_expectations('sha1abc', 'sha1abc', true)
-    record_expecations_after_build
+    record_expectations_after_build
+
+    PmdTester::PmdReportBuilder
+      .new(projects, options, options.base_config, options.base_branch)
+      .build
+  end
+
+  # In CI, there is no previous existing distro that can be reused,
+  # that means target/pmd-bin-6.10.0-SNAPSHOT-master-sha1abc does not
+  # exist. However, pmd-dist/target/pmd-bin-6.10.0-SNAPSHOT.zip exists
+  # from a previous build and should be reused.
+  def test_build_skip_ci
+    projects = []
+    argv = %w[-r target/repositories/pmd -b master -p pmd_releases/6.1.0
+              -c config/design.xml -l test/resources/project-test.xml]
+    options = PmdTester::Options.new(argv)
+
+    FileUtils.mkdir_p 'target/repositories/pmd/pmd-dist/target'
+    FileUtils.touch "target/repositories/pmd/pmd-dist/target/pmd-bin-#{@pmd_version}.zip"
+
+    record_expectations('sha1abc', 'sha1abc', false)
+    PmdTester::Cmd.stubs(:execute).with("unzip -qo pmd-dist/target/pmd-bin-#{@pmd_version}.zip" \
+      ' -d pmd-dist/target/exploded').once
+    PmdTester::Cmd.stubs(:execute).with("mv pmd-dist/target/exploded/pmd-bin-#{@pmd_version}" \
+      " #{Dir.getwd}/target/pmd-bin-#{@pmd_version}-master-sha1abc").once
+    record_expectations_after_build
 
     PmdTester::PmdReportBuilder
       .new(projects, options, options.base_config, options.base_branch)
@@ -24,13 +67,15 @@ class TestPmdReportBuilder < Test::Unit::TestCase
               -c config/design.xml -l test/resources/project-test.xml]
     options = PmdTester::Options.new(argv)
 
-    # File does not exist yet this time...
+    # PMD binary does not exist yet this time...
     record_expectations('sha1abc', 'sha1abc', false)
-    PmdTester::Cmd.stubs(:execute).with('git checkout master')
-                  .returns('checked out branch master').once
     PmdTester::Cmd.stubs(:execute).with('./mvnw clean package -Dmaven.test.skip=true' \
                   ' -Dmaven.javadoc.skip=true -Dmaven.source.skip=true -Dcheckstyle.skip=true').once
-    record_expecations_after_build
+    PmdTester::Cmd.stubs(:execute).with("unzip -qo pmd-dist/target/pmd-bin-#{@pmd_version}.zip" \
+                  ' -d pmd-dist/target/exploded').once
+    PmdTester::Cmd.stubs(:execute).with("mv pmd-dist/target/exploded/pmd-bin-#{@pmd_version}" \
+                  " #{Dir.getwd}/target/pmd-bin-#{@pmd_version}-master-sha1abc").once
+    record_expectations_after_build
 
     PmdTester::PmdReportBuilder
       .new(projects, options, options.base_config, options.base_branch)
@@ -48,8 +93,8 @@ class TestPmdReportBuilder < Test::Unit::TestCase
 
     projects[0].auxclasspath = '-auxclasspath extra:dirs'
     record_expectations('sha1abc', 'sha1abc', true)
-    record_expecations_after_build
-    record_expectations_project_build
+    record_expectations_after_build
+    record_expectations_project_build('sha1abc')
 
     PmdTester::PmdReportBuilder
       .new(projects, options, options.base_config, options.base_branch)
@@ -71,8 +116,8 @@ class TestPmdReportBuilder < Test::Unit::TestCase
 
     projects[0].auxclasspath = '-auxclasspath extra:dirs'
     record_expectations('sha1abc', 'sha1abc', true)
-    record_expecations_after_build
-    record_expectations_project_build(true)
+    record_expectations_after_build
+    record_expectations_project_build('sha1abc', true)
 
     PmdTester::PmdReportBuilder
       .new(projects, options, options.base_config, options.base_branch)
@@ -81,45 +126,49 @@ class TestPmdReportBuilder < Test::Unit::TestCase
 
   private
 
-  def record_expectations_project_build(error = false)
+  def record_expectations_project_build(sha1, error = false)
     PmdTester::ProjectBuilder.any_instance.stubs(:clone_projects).once
     PmdTester::ProjectBuilder.any_instance.stubs(:build_projects).once
     PmdTester::SimpleProgressLogger.any_instance.stubs(:start).once
     PmdTester::SimpleProgressLogger.any_instance.stubs(:stop).once
-    File.stubs(:exist?).with('target/reports/master/checkstyle/pmd_report.xml').returns(false).once
     error_prefix = error ? 'PMD_JAVA_OPTS="-Dpmd.error_recovery -ea" ' : ''
+    distro_path = "#{Dir.getwd}/target/pmd-bin-#{@pmd_version}-master-#{sha1}"
     PmdTester::Cmd.stubs(:execute)
                   .with("#{error_prefix}" \
-                        'target/pmd-bin-6.10.0-SNAPSHOT/bin/run.sh ' \
+                        "#{distro_path}/bin/run.sh " \
                         'pmd -d target/repositories/checkstyle -f xml ' \
                         '-R target/reports/master/checkstyle/config.xml ' \
                         '-r target/reports/master/checkstyle/pmd_report.xml ' \
                         '-failOnViolation false -t 1 ' \
                         '-auxclasspath extra:dirs').once
-    Dir.unstub(:getwd)
-    Dir.stubs(:getwd).returns('current-dir').twice
     PmdTester::PmdReportDetail.any_instance.stubs(:save).once
   end
 
   def record_expectations(sha1_head, sha1_base, zip_file_exists)
-    Dir.stubs(:getwd).returns('current-dir').once
-    Dir.stubs(:chdir).with('target/repositories/pmd').yields.once
-    PmdTester::Cmd.stubs(:execute).with('git rev-parse HEAD').returns(sha1_head).once
-    PmdTester::Cmd.stubs(:execute).with('git rev-parse master').returns(sha1_base).once
+    Dir.expects(:chdir).with('target/repositories/pmd').yields.once
+    PmdTester::Cmd.stubs(:execute).with('git rev-parse master^{commit}').returns(sha1_base).once
+    # inside checkout_build_branch
+    PmdTester::Cmd.stubs(:execute).with('git checkout master')
+                  .returns('checked out branch master').once
     PmdTester::Cmd.stubs(:execute).with('./mvnw -q -Dexec.executable="echo" ' \
                   "-Dexec.args='${project.version}' " \
                   '--non-recursive org.codehaus.mojo:exec-maven-plugin:1.5.0:exec')
-                  .returns('6.10.0-SNAPSHOT').at_least(1).at_most(2)
-    # File does not exist yet this time...
-    File.stubs(:exist?).with('pmd-dist/target/pmd-bin-6.10.0-SNAPSHOT.zip')
-        .returns(zip_file_exists).once
+                  .returns(@pmd_version).at_least(1).at_most(2)
+    PmdTester::Cmd.stubs(:execute).with('git status --porcelain').returns('').once
+
+    # back into get_pmd_binary_file
+    PmdTester::Cmd.stubs(:execute).with('git rev-parse HEAD^{commit}').returns(sha1_head).once
+    # PMD binary might not exist yet...
+    distro_path = "target/pmd-bin-#{@pmd_version}-master-#{sha1_base}"
+    if zip_file_exists
+      FileUtils.mkdir_p(distro_path)
+    elsif File.exist?(distro_path)
+      Dir.rmdir(distro_path)
+    end
   end
 
-  def record_expecations_after_build
-    PmdTester::Cmd.stubs(:execute).with('git rev-parse HEAD').returns('sha1abc').once
+  def record_expectations_after_build
     PmdTester::Cmd.stubs(:execute).with('git log -1 --pretty=%B').returns('the commit message').once
-    PmdTester::Cmd.stubs(:execute).with('unzip -qo pmd-dist/target/pmd-bin-6.10.0-SNAPSHOT.zip' \
-                  ' -d current-dir/target').once
     PmdTester::PmdBranchDetail.any_instance.stubs(:save).once
     FileUtils.stubs(:cp).with('config/design.xml', 'target/reports/master/config.xml').once
   end